can fire bullet with joystick and mouse

+++ Assets/Scripts/ProjectileController.cs
@@ -4,34 +4,41 @@ using System.Collections;
 public class ProjectileController : MonoBehaviour {
 
   private float selfDestroyTimer = 1;
+  private float bulletSpeed = 10.0f;
+  private Rigidbody2D bulletBody;
+  private string bulletOwner;
+  private bool bulletFiring = false;
   // Use this for initialization
   void Start() {
-
+    bulletBody = GetComponent<Rigidbody2D>();
   }
 
   // Update is called once per frame
   void Update() {
-    Vector3 position = this.transform.position; // testing purposes only
-    position.y--;						// testing purposes only
-    this.transform.position = position;
+    if (bulletFiring) {
+      executeBulletFire();
+    }
+  }
 
+  void OnTriggerEnter2D(Collider2D col) {
+    if (col.gameObject.tag == "Player" && col.gameObject.name != bulletOwner && bulletFiring) {
+      col.transform.gameObject.GetComponent<PlayerController>().Kill(); //call player kill function
+      Destroy(gameObject);
+    }
+  }
+
+  public void SetOwnerAndShoot(string owner) {
+    bulletOwner = owner;
+    bulletFiring = true;
+  }
+
+  private void executeBulletFire() {
+    bulletBody.AddForce(transform.up * -bulletSpeed);
     selfDestroyTimer -= Time.deltaTime;
 
     if (selfDestroyTimer <= 0) {
       Destroy(gameObject);
     }
   }
-
-  void OnCollisionEnter2D(Collision2D col) {
-    if (col.gameObject.tag == "Player") {
-      Debug.Log("destroyed player");
-      col.transform.gameObject.GetComponent<PlayerController>().Kill(); //call player kill function
-      Destroy(gameObject);
-    }
-  }
-  public void die() {
-    Debug.Log("Die");
-  }
-
 
 }
